move `Program`-wide stuff into `chalk::program`
